video_processing/requirements.txt
pytube
tqdm
openai-whisper